+++ src/neural_network.lua
@@ -55,6 +55,7 @@ function ann:initw()
   local epsilonOUT = math.sqrt(6) / math.sqrt(self.Lhid + self.Lout)
   --
   local function initmat(din, dout, value)
+    math.randomseed(os.time())
     local mat = {}
     for i = 1, dout do
       local aux = {}
@@ -72,10 +73,10 @@ end
 
 
 -- PRE:
--- input - column [N,1] vector (table).
+-- input - feat [1,N] vector (table).
 --
 -- POST:
--- Returns output [K,1] vector (table).
+-- Returns output [1,K] vector (table).
 
 function ann:predict(input)
   local function matprod(m1, m2)
@@ -105,10 +106,11 @@ function ann:predict(input)
     local y = 1 / (1 + math.exp(-x))
     return y
   end
+  -- input must be a column [N,1] vector (table).
   -- step 1
   local aIN = {{1}}
   for i = 1, #input do
-    table.insert(aIN, {input[i][1]})
+    table.insert(aIN, {input[i]})
   end
   -- step 2
   local zHID = matprod(self.thetain, aIN)
@@ -121,6 +123,92 @@ function ann:predict(input)
   for i = 1, #azOUT do
     azOUT[i][1] = sigmoid(azOUT[i][1])
   end
-  return azOUT
+  local flatOUT = {}
+  for i = 1, #azOUT do
+    table.insert(flatOUT, azOUT[i][1])
+  end
+  return flatOUT
+end
+
+
+-- PRE:
+-- feat - list of example feature vectors (table).
+-- targ - list of target value vectors (table).
+--
+-- POST:
+-- Fits the neural network params to the given data.
+-- Returns training error (number).
+
+function ann:train(feat, targ)
+  require("gradient_ascent")
+  local function saveThetas()
+    local thetas = {}
+    -- theta in
+    for i = 1, self.Lhid do
+      for j = 1, (self.Lin + 1) do
+        table.insert(thetas, self.thetain[i][j])
+      end
+    end
+    -- theta out
+    for i = 1, self.Lout do
+      for j = 1, (self.Lhid + 1) do
+        table.insert(thetas, self.thetaout[i][j])
+      end
+    end
+    return thetas
+  end
+  local function loadThetas(thetas)
+    -- theta in
+    local index = 1
+    for i = 1, self.Lhid do
+      for j = 1, (self.Lin + 1) do
+        self.thetain[i][j] = thetas[index]
+        index = index + 1
+      end
+    end
+    -- theta out
+    for i = 1, self.Lout do
+      for j = 1, (self.Lhid + 1) do
+        self.thetaout[i][j] = thetas[index]
+        index = index + 1
+      end
+    end
+  end
+  local function cost(thetas)
+    local sqerr, pr
+    local J = 0
+    loadThetas(thetas)
+    for m = 1, #feat do
+      pr = self:predict(feat[m])
+      sqerr = 0
+      for k = 1, #pr do
+        sqerr = sqerr + math.pow(targ[m][k] - pr[k], 2)
+      end
+      J = J + sqerr
+    end
+    J = J / #feat
+    -- Regularisation
+    local R = 0
+    for i = 1, #self.thetain do
+      for j = 2, #self.thetain[1] do
+        R = R + math.pow(self.thetain[i][j], 2)
+      end
+    end
+    for i = 1, #self.thetaout do
+      for j = 2, #self.thetaout[1] do
+        R = R + math.pow(self.thetaout[i][j], 2)
+      end
+    end
+    R = R * (0.01 / (self.Lin*self.Lhid + self.Lhid*self.Lout))
+    return -(J + R)
+  end
+  -- flatten thetas
+  local flatTheta = saveThetas()
+  gradient_ascent(cost, flatTheta, 0.1)
+  -- deflat theta, restore model
+  trerr = cost(flatTheta)
+  loadThetas(flatTheta)
+  -- return training err
+  return trerr
 end